Create Employee with most years of service
This file reads in a list of employees first names, last names, time in service, and the title held. 
It also compares the years of service and produces the team member with the most years of service using a While Loop and an if statement ( hightime < time)
then hightime = time. ----> this way, the highest time comes out at the end. Also, the first name and last name and title with the highest time is tracked