Update update.php
Update sends to the database ... No need to Filter_SANITIZE_INT
This form takes all the fields from Edit.php and sends the fields to the update.php file. The variables are carried over and filtered. 
Then the database is updated and we prepare the query and echo out the results